+++ src/main/java/org/cardanofoundation/rewards/entity/Reward.java
@@ -2,12 +2,14 @@ package org.cardanofoundation.rewards.entity;
 
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class Reward {
-    private double amount;
+    private BigInteger amount;
     private String stakeAddress;
 }
