+++ src/main/java/org/cardanofoundation/rewards/entity/jpa/DbSyncReward.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import org.hibernate.annotations.Immutable;
 import org.springframework.context.annotation.Profile;
 
+import java.math.BigInteger;
+
 @Entity
 @Immutable
 @Getter
@@ -21,7 +23,7 @@ public class DbSyncReward {
     @EqualsAndHashCode.Exclude
     private DbSyncPoolHash pool;
 
-    private Double amount;
+    private BigInteger amount;
     private String type;
 
     @ManyToOne
