+++ src/main/java/org/cardanofoundation/rewards/repository/DbSyncTransactionRepository.java
@@ -5,12 +5,14 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
 @Profile("db-sync")
 public interface DbSyncTransactionRepository extends ReadOnlyRepository<DbSyncTransaction, Long> {
     @Query("SELECT SUM(deposit) FROM DbSyncTransaction WHERE block.epochNo = :epoch")
-    Double getSumOfDepositsInEpoch(Integer epoch);
+    BigInteger getSumOfDepositsInEpoch(Integer epoch);
 
     @Query("SELECT SUM(fee) FROM DbSyncTransaction WHERE block.epochNo = :epoch")
-    Double getSumOfFeesInEpoch(Integer epoch);
+    BigInteger getSumOfFeesInEpoch(Integer epoch);
 }
